Correct Frontmatter#published? default, which returned an empty text
The published? method was returning an empty text as the default
value when published? was not defined. This commit fixes the bug by
returning a boolean instead.

Since some of the imported posts don't have a published field defined
in their frontmatter, the default boolean value to be returned by this
method is set to true.