Create a class to get all people from the current term of a legislature.
This class copies the #people method from the TermTable page in
viewer-sinatra, so that all people from the current term of a specific
legislature can be displayed in the corresponding page for that
legislature.

    https://github.com/everypolitician/viewer-sinatra/blob/4c12a77195d3f000fa220bf2b9e6ded06e6d767e/lib/page/term_table.rb#L58

Since we are not going to reuse the code from the TermTable and
we also don't need all the code there, this class will be coded to
respond to the particular needs of this site.

The #people method from TermTable was renamed to #find_all.

It also uses our mapit wrapper object since an EP::Person needs it.

The functionality in this class could be useful for the following pages,
which make use of it to list people in a legislature:
* /position/senator/
* /position/representative/
* /position/executive-governor/

There is a #none? method that checks if there are no persons
with an id in a legislature. This method can be used in the routes
to jump to the next similar route and check if the id exists in other
legislatures.