Add a view for a person page
The route uses the previously created Page::Person presenter to
send the available information about the person.

The code for the layout in this view was taken from the Jekyll version
of Shine Your Eye. In particular:

*  shineyoureye/_layouts/people.html

    https://github.com/theyworkforyou/shineyoureye/blob/fd313124f525782fa11eabc47f353ac9f70c8873/_layouts/people.html

The id used to find a single person comes from the url slug of a
person page. Ideally this will be Pombola url slugs that will be
mapped to EP UUIDs. For now, since we're still working on getting
that data into EveryPolitician, the find_single method class merrily
uses the id as an EP UUID.

In this app, if no such id can be found for the house of
representatives, we tell Sinatra to pass and jump to the next route that
is the same, so that we can search for the same id in the Senate or
governors.

If there are no more matching routes or the id is not found anywhere,
a 404 error will be thrown.