Add a presenter page for the person page view
This page works as a presenter for a person page. The person object
is obtained from PeopleByLegislature.

The person view asks for a summary, executive_positions, job_history
and education. Since these are not implemented yet, but are part of
the partner's requirements, they are left as placeholders.

The person view used as a guide for this page can be found in the Jekyll
version of Shine Your Eye:

*  shineyoureye/_layouts/people.html

    https://github.com/theyworkforyou/shineyoureye/blob/
    fd313124f525782fa11eabc47f353ac9f70c8873/_layouts/people.html

The urls point to the same pages as the Pombola version of the site:
    http://www.shineyoureye.org/person/sylvester-o-ogbaga/