Refactor the method to create a tempfile in the test_helper
We are using a block so that the file is closed implicitly right after it's
used even if there is a exception.

Also we are now returning the filename directly instead of the tempfile
object.

The MarkdownWithFrontmatter tests were updated to make use of the
new signature (they are the only tests affected).