Add a presenter page for the posts view
This presenter obtains all post filenames from the specified posts
directory and returns a list of MarkdownWithFrontmatter objects
with the baseurl set to the right path for blog posts.

The posts are ordered by reverse date.

It also has a method with the logic to format the date of a post
displayed in the list.