Add a breadcrumb builder class to build the list of breadcrumbs
This class builds the list of breadcrumbs for the current page,
using the url of said page to build the routes to be used as urls for
each breadcrumb.

The path of the current page (the last breadcrumb) is excluded and
no breadcrumb is returned for it. This is because the last breadcrumb
is created manually in the breadcrumbs view, using the current page
title as display text and no link.

Not hyperlinking the last breadcrumb is a best practice as the text
leads to the same page, so it would be confusing for the user.