Add class to isolate all knowledge of the routes-to-display data structure
For the breadcrumbs, the route-to-display-text data structure/map is used
by people to define their routes there, together with the display text they
want them to show in the breadcrumbs view.

This data structure can be changed by users to accomodate their needs.
Hence this wrapper to isolate the application from changes to the data
structure.