Let the deploy script use a specific version of content from prose.io
The deploy script can now be passed an argument so that we can
git checkout a specific commit-ish in the content repository after
cloning it, and before symlinking the media files.

This is done so that the tests can be run at a specific point in the
history of the prose repository, in order to get predictable test output.