Move the knowledge about the route from the posts page to the app
At the beginning it looked like a good idea that a page would know
about it's baseurl, however this is actually spreading the knowledge
about the route, so if the route needs to be changed in the future,
there are two different places where we would have to update it.

This commit moves that knowledge so that it stays in the place where
the routes are defined.