Add a finder class to perform search operations on disk
There was some repetition in the posts, post and info pages that
suggested that there was another class hiding inside of them. The
repetition was extracted into a class that searches for documents
with a certain name in a certain directory.

By doing this the dependency on `MarkdownWithFrontmatter` is now
in one single class (the `Finder` class) instead of spread across all
page classes. Changes to `MarkdownWithFrontmatter` affect only the
`Finder` class.