Use the finder class in the posts, post and info pages.
There was some repetition in the posts, post and info pages that
suggested that there was another class hiding inside of them. The
repetition was extracted into a class that searches for documents
with a certain name in a certain directory.

By doing this the dependency on `MarkdownWithFrontmatter` is
now in one single class (the `Finder` class) instead of spread
across all page classes. Changes to `MarkdownWithFrontmatter`
affect only the `Finder` class.

As a consequence, now the presenters contain only the logic for
the views and no extra dependencies. They depend on an
abstraction that responds to the relevant messages.

The site-specific definitions, in particular those related to filepaths
and filenames, are now in one place (the constants helper), and
easy to edit per country.

As an extra consequence, the repetition in the page tests was
eliminated as well, and the tests are now much simpler.