Add a title field to the posts page
For the moment, and until different requirements come in for either
the post pages or the event pages, it seems that the events page
can use the same page as the posts.

The only change that must be introduced is a title field, which is
for the moment the only distinction between the two.

Since the title is site dependent, I decided to pass it as an
argument, so that it has to be defined in the app.rb