Add a presenter page for the homepage
For now the homepage doesn't contain logic for the featured persons,
since the logic for that is not ready yet. In the meanwhile, the featured
persons will be hardcoded in the homepage view.

Also, since we still don't have an event_date method in documents
nor an evendate field defined in prose.io config file, the featured_events
and the featured_posts work in the same way.