Fix the title in the head partial
The meta title of a page (the one that goes inside the head tag) is
different for the homepage and the rest of pages. The Jekyll code
that was initially taken from the Jekyll site was updated to
reproduce the behaviour of the homepage meta title.

In order for this fix to work, all normal pages need to have a title.

The only page without a title was the /places/ route. This route was
deleted since it was a placeholder. The correct route is /places
and it is already covered by the info pages.