Update the homepage with the code from the Jekyll site
The homepage is mostly harcoded for now, in particular the featured
persons. The homepage view code was taken from:

*  shineyoureye/index.html

    https://github.com/theyworkforyou/shineyoureye/blob/
    8441907b3d179f5cabb828ecebd1b904cb98ad89/index.html

There are no tests for the home page view because the homepage
as it is in the Jekyll version of the site is the whole HTML document,
i.e. it is not generated using the layout. So some parts in the home
page layout are different to the rest of the pages' layout (probably
changes were made in the generic layout that all pages use and
the home page got out of sync).

I identified the common parts and loaded the different ones using
the homepage checker that was added to the layout helper. Since
the only differences seem to be styles and markup, I think these
two different branches of the if statement can be eliminated by
the designers, unifying styles and markup for both versions.

The parts where there is a conditional check are:

* the part where the styles are imported in the head tag (_head partial)

* the styles in the footer (_footer partial)

* a section in the body (layout view)

In my opinion the section in the body can stay, but the sections in
the head and footer could easily be simplified by the designers to
unify the markup and the css. That would allow the conditional to
go away.