Add an event date to the frontmatter parser
An eventdate field is used in the prose configuration so that event
documents can be saved with an eventdate field in the frontmatter.

This field can be used to select events to display in the home page.

This commit adds a convenience method to retrieve the field and
provide a default if the field is not present.