Add an event_date method to the MarkdownWithFrontmatter class
After the event_date was added to the frontmatter parser, this
class was updated to reflect the change.

If no event date can be found from the frontmatter, the new
event_data method returns nil

Tests were added to document the fact that we can parse a variety
of date formats, which is helpful since the field in prose isn't
constrained to be a date.

This might feel a bit like we're writing test for library code, which
is generally not worth it, but someone might well change the
implementation of event_date in the future to parse the date in
some less flexible way, and it would be nice for the tests to catch
a loss of useful functionality if that happens.