Update home page to select upcoming events and display them
The events are selected by event day. Only the events happening
in the future are shown in the homepage. This includes events
happening "today".

The events are sorted from those happening sooner to those
happening later. From those, the three closest ones are taken.

If an event doesn't have an event date, an error is thrown.