Add basic page to hold the idea that every page should have a title
All presenter pages should have at least a title. If there isn't any
other logic in the view that needs to be handled by a presenter, but
they need to use the site's layout this page can be used instead.

For example, the Jinja2 template or the search page are examples of
pages we are using at the moment that don't really need any logic in
the view, but still need a title.

This allows to avoid having to do checks on objects to see if they
respond to messages (the title, in this case) which is a code smell,
and also allows to define the value of those messages in a
single place. For pages that don't use the site's layout (for example
the redirect view) this is not relevant.