Add basic breadcrumbs functionality
This provides functionality for breaking down the path part of a URL
an provide a list of links to the information hierarchy represented by
the /-separation, and linking to pages at the higher levels. (These
links at the top of the page are often referred to as "breadcrumbs".)

For example, you might configure the following prefix map:

  {
    '/' => 'Homepage',
    '/politicians', => 'People',
    '/politicians/mps', => 'MPs',
  }

... and then the following breadcrumbs HTML would be generated for the
path '/politicians/mps/':

  <ol class="breadcrumb">
    <li><a href="/">Homepage</a></li>
    <li><a href="/politicians">People</a></li>
    <li class="active">MPs</li>
  </ol>

Sometimes intermediate breadcrumbs won't have anywhere sensible to link
to - for example, while info pages are all of the form
'/info/hello-world', there is no index of info pages at '/info/', so you
can skip any prefix from producing a breadcrumb by mapping it to nil.
For example, you could use the prefix map:

  { '/info' => nil }

... so that the breadcrumb that would otherwise link to '/info' would be
omitted.