Update countries fixture, since Assembly was renamed
Assembly was renamed to House of Representatives in commit
ccdc6d1:

    https://github.com/everypolitician/everypolitician-data/commit/
    ccdc6d143f1bceae8b3d99d853269cdf2dcc82a8

Because of this, we have to update the countries fixture so that web
tests still work when trying to get the House of Representatives
through legislature('Representatives') in the app.

A trimmed version was added this time, since the only country that
is relevant to this site is Nigeria.

This is the file at commit d8a4682:

    https://github.com/everypolitician/everypolitician-data/commit/
    d8a4682f9615ac69429eeeb5a9758037753c29c4

Note:
In the older version of the file, the commit id in the popolo_url of
the countries json file was trimmed to seven characters.The new
version shows the full commit id. For convenience though, the
commit id of the url to the popolo file in this new version of the
countries fixture was trimmed to seven characters so that we
can use the stub method with a short commit id.