Add a presenter page for a list of places/areas
Places is a generic page that can be used for all the submenu
pages that list states, districts, areas, etc. It can be used for
the three subpages under Places.

As described in the requirements document, the three places
pages to implement are:

- States Page
A page with a list of all the states on it. [...]

- Senatorial Districts Page
Same frame as the states page but for senatorial districts.

- Federal Constituencies Page
Same frame as the states page but for federal constituencies.

For now it is just a struct to pass the view what it needs: the
title and the areas.