Add a view for the page that shows a list of constituencies
There was no view for this page in the Jekyll version of the site,
so the code for this view is a mix of

- The general structure found at the pombola version of the site:
    http://www.shineyoureye.org/place/is/federal-constituency/hr2011

- The relevant Bootstrap components for the different parts of the
page. The original code for those parts was replaced with a Bootstrap
component:
    http://getbootstrap.com/components/

In particular, the different parts of this page are:

- A list of links to all the possible places that can be displayed.
Bootstrap component used: http://getbootstrap.com/components/#nav-pills
Comments: The pombola page has pages for most of those links, but
according to our requirements, we seem to have to implement only
three of those. So for now, this section is commented.

- The list of federal constituencies.
Bootstrap component used: http://getbootstrap.com/components/#media-list
Comments: There is a default image rendered. The route to that image is
"/static/images/place-90x90.d4bd412d530a.jpg" in the pombola version of
the site. We don't have any file named like that but we have
"/images/place-200x200.jpg", so I used that one instead.

- The pagination.
Bootstrap component used: http://getbootstrap.com/components/#pagination
Comments: Since pagination is not implemented yet, this section was
commented.