Promote the checker methods to be public
Since this code could be used outside of our particular application
it is a good practice to make the code robust by dealing with errors
where they may happen, for example calling first on something
that could be nil.

However, it can be also useful to provide the checker methods
none? and multiple? as part of the API, so that devs can choose
if they want to:

  1) check first that they can call the method safely or
  2) just call the method and handle the exception in some way

This will hopefully make this class more flexible and usable.