Pass if no file is found in the routes that are dynamic
Of the two ways the finder offers to handle an error when
searching for a document, the checker-methods way was
choosen, to avoid throwing an exception if there is an error
when calling "find_single".

If this happens, we just pass and let Sinatra throw its generic
Sinatra::NotFound error.