First stab at a place helper class
For now this class contains only the functionality needed for the
current pull request this commit is part of.

It uses a mapit area to get the name of the place. If it is a
federal cosntituency for example, it expects the area to have
a state name field. This field does not exist in the data
returned by mapit, so this class uses an extended area.
A wrapper class will be used to create this extended area
by adding this field so that we can use it in the presenter.

The presenter will return a list of place objects for each area.

We are using the same slugs as the Pombola version of the site,
using a mapping between mapit ids and Pombola slugs