Add mapit-to-EveryPolitician (EP) area-ids mapping csv file
This file was generated using a python snippet which uses
the Levenshtein algorithm to check if two area names
correspond to the same area:

    https://github.com/mysociety/pombola/blob/master/bin/name_matcher.py

This snippet takes a csv file as input, containing two columns of
unrelated area names, and returns an output csv file with those
names matched. The 100% matches are in the first rows, then
similar matches, then names with no matches.

The csv file added in this commit contains two columns, the left
one are mapit ids and the right one are the corresponding EP
ids. The process to generate it was the following:

* Fetch mapit areas from:
    http://www.shineyoureye.org/mapit/areas/FED

* Fetch EP areas from:
    https://cdn.rawgit.com/everypolitician/everypolitician-data/d5a1eb71c6ce0172752fcd081b0fb6d5f975c61e/data/Nigeria/Representatives/ep-popolo-v1.0.json

* Generate input csv for the name matcher, containing mapit
area names in the left column and EP area names in the right
column

* Generate output csv, feeding the previous csv file to the
name-matcher snippet

* Perform a reconciliation of that output file, showing the two
area names and asking a human if they are the same area

* If they are the same area, get the ids of those names and
generate a csv file with the mapit ids in the left column and
the corresponding EP ids in the right column (THIS FILE).

* If they are not the same area, generate a csv file with the
name of the mapit area in the left column and the name of
the wrongly assigned EP area for later manual treatment.