Add a wrapper class to deal with mapit data
This class gets the federal constituencies area information as well
as the states area information from the mapit API, and parses it
so that it can be used by the presenter classes.

In particular, it adds the state corresponding to a federal
constituency to its area information, because it's needed to be
displayed in the list of areas pages. This is done because at
the moment, the 'parent_area' field comes with a null value
from the mapit API.

In order to associate each constituency with its state, a mapping
file is used. This mapping file contains area ids in the left column
and their corresponding state id in the right column.

If this was changed in the future, for example making mapit return
the id of the corresponding state the constituency is part of, then
most of the code in the private section could be removed.

It also uses a mapping file to pair mapit ids to their Pombola slugs
so that the places are created with the same urls they have in
Pombola.