Create a Person class based on the PersonCard in viewer-sinatra
Since it looks like we are not going to be able to share functionality
with viewer-sinatra, this Person class was created as a simplified
version of the PersonCard in viewer-sinatra:

    https://github.com/everypolitician/viewer-sinatra/blob/595f5b6e018416b9ea0de19669fc742bc0e78897/lib/person_cards.rb

This class decorates an EP person object using Forwardable to
delegate only the methods we want to expose from it.

The methods were added according to the requirements document.
For now it has just the functionality to match the requirements for a
representative, as long as that functionality is also provided by the
EP gem (for example, the job history is not provided by EP).

It also uses our mapit wrapper object so that it can map EP area
ids to Mapit area ids and return a Mapit area when asked for its
area.

Since some of this data is not available in EP for the specific case
of Nigeria (as is the case with social media, emails, etc), the tests
extend the person instance with fake data for now, based on
the contents of those fields that EP would return, so that the
existence of this functionality is documented, even if the data is
not available yet for Nigeria. When this data is not available, the
EP gem returns a nil. Asserting for nil in tests would not make it
clear if it is because the data is missing or because EP doesn't
provide that field.

We also need some of the extensions to the LegislativePeriod class
of the EP gem that were added to viewer-sinatra.

However, we don't need them all. This commit adds the extension
module with just the functionality that we need, which is the #people
method.

The code for the LegislativePeriodExtension was taken from:

    https://github.com/everypolitician/viewer-sinatra/blob/881038c92a86783efabf68b525ad1e6c9f9d1bf3/lib/everypolitician_extensions.rb

From the requirements document:
Representatives pages must have:

* Politician's Name
* Constituency they represent
* Date of Birth
* Political Party they are a member of
* Title (eg Senator, Representative, Governor)
* Committee Membership details
* Social Media details (twitter and facebook)
* Email addresses
* NASS office allocations (the office number in the national assembly)
* Constituency Office Address
* Phone number
* Educational and Training Information
* Job history
* Notes tab (nice to have)
* Summary field for SYE to put in some free text information about the politician.
* Picture of the politician
* Facebook comments section
* Site and PU search box (just one box that does both things)