Add a method in the finder to find optional documents
Summaries are optional, blog posts are not, so when trying to use
find_single to find a summary, we have these scenarios:

* If we just pass if finder.none?, that would make summaries
mandatory as we will get a 404 if no summary is found.

* If we don't use none?, then a 404 is raised anyway, through
a custom exception, if no summary is found.

So this commit adds a method that allows to have summaries
as optional documents. Instead of `find_single` we can now
call `find_or_empty`, and if no summary is found it will
just return a duck type that responds to anything we need it
to respond to, kind of like a null object.

Then in the view, nothing will be shown if there is no summary.

This also prevents us from having to check for nil everytime we
search for a summary.