Add a place view and its route
Since there is no template for a place page in the Jekyll version of
the site, the markup of the Pombola version was used instead:

    http://www.shineyoureye.org/place/abakalikiizzi/

Then, the different sections of the page where replaced by
Bootstrap components:

* For the main sections, the same markup as that of a person
page was used. That's why some classes are "person__image"
or "person__name".

* For the links to people and places, the "Pills" component in
Bootstrap was used:

    http://getbootstrap.com/components/#nav-pills

The sections in the Pombola site for which we still don't have the
functionality implemented were left commented.