Add an empty page, usuable as a Jinja2 template
Interactive elements of the site have to be provided by other
applications, usually hosted at subdomains of the main site.
A requirement for these is that they have the same look and feel as the
main static site, but we don't want to have to maintain HTML, CSS and
Javascript for them in parallel. Instead, we can serve a template from
the static site, like any other page. This commit adds one which is
suitable for Jinja2, the default templating language of Flask (which
we're writing the polling unit search lookup in).