Add a route to list all states
At the moment, the Pombola version of the site shows some
information on place items in the places list when they are child
areas, but not when they are parent areas.

* Parent areas:
    http://www.shineyoureye.org/place/is/state/

* Child areas:
    http://www.shineyoureye.org/place/is/federal-constituency/hr2011

So the places view was updated to show that.

Also, Everypolitician does not give us governors yet, so for now
there is a governors method to instantiate the EP people but it
just returns nil. This doesn't matter at the moment since states
don't show legislature information as commented above.