Add a script to incorporate personal history from Pombola
This is a quick and hacky script to take person summaries and
past positions from the Pombola-based ShineYourEye and turn them into
markdown-with-frontmatter files for prose.io under the 'summaries'
directory. These are incorporated into the person page for display.

This skips party memberships, since they're likely to be less accurate
than that from EveryPolitician or any memership where the organization
or role are missing.