Only load MapIt data once per server instance
Mapit::Wrapper was being instantiated on every request that needed MapIt
data, and so new requests to the MapIt API were being made on every request
that required them. This is a waste of resources, and massively slows
down crawling the site - the fix in this commit speeds up the crawling
of the site from about 10 minutes to about 2 minutes on my laptop.

There's no advantage in 'governors', 'representatives' and 'senators'
being functions, or from 'mapit_url' and 'mappings' being helpers
(they're only used in app.rb) so these are just variables in app.rb now.
This also makes it more explicit where that data comes from.