Remove area type specific code from Mapit::Mappings
The parameters to the initializer and one of the method names of
Mapit::Mappings include the area type, and limited the number of mapping
filenames used for parent-child relationships and MapIt <-> EP area IDs
mappings. There's no need for this - the class can be more generic in
its handling of these CSV files, which will make it more easily
extendable in this app and reusable for others.