Add tests for the content of the 'unlinked' pages
There are two pages that aren't linked from the homepage - a special
page for the crawler to start from, and a page that just serves up a
Jinja2 template for the Polling Unit search lookup app to use.

This commit adds tests that check that these pages exist and have the
content that's important on them.