Use all person methods that are exposed in EP person
We were exposing some methods in person but not using them in
the view. This commit adds those and also tests for the parts of the
view that were untested.

The Facebook tests are commented because no senator or
representative has a facebook account, but governors do, so then
we will be able to fetch a governor and uncomment them.