Add a person class to deal with person data coming from Morph
Some fields like twitter or facebook come with several values
separated by semicolons. To keep the same behavior as the EP
person class, I am taking the first of those values.

For the same reason, when the phone comes with several values,
all of them are returned separated by commas as a string.

This class adds a slug method, to be used in the route when we
want to find a specific governor.

Also. a wikipedia_url method was added although there is no
wikipedia url coming from morph, because this method is part
of a person API.