Remove hardcoded area types from Mapit::Wrapper
The Mapit::Wrapper class hardcoded particular types of area in
method names like 'federal_constituencies' or 'states'. This means you
had to change the code of the class to reuse it for other area types,
which is undesirable. This commit changes that so that there is a
'places_of_type' method which returns all the places for a
particular MapIt type code, e.g. '.places_of_type('SEN')'

Unforunately, because of the way that parent areas are set with this
version of the code, it's necessary to distinguish the area type that
parents may be drawn from and area types that child areas are drawn
from. This will be made more flexible in the next commit, however.