Simplify Mapit::Wrapper and make it more efficient
This class had a number of problems which this commit fixes:

 - It was very inefficient, iterating over all areas whenever it need to
   look one up.

 - It dealt with heterogenous representations of areas (MapIt-derived
   hashes and Mapit::Place objects) even after the wrapper object has
   been constructed.

 - The parent-child relationships had to be set very carefully, in that
   the parent places had to be created before the child places, so that
   the parents would exist before setting the parent attributes of the
   children.

This commit fixes those problems and simplifies the class both to use
and reason about considerably.

Setting the parent of an Place is now done either by passing in a parent
Place object when initializing the Place, or by setting it afterwards
with the place= setter.