Switch to using the local copy of the MapIt Area data
To avoid having to fetch data from MapIt every time the server is
started, we can use the area JSON in the repository. (The MapIt
data changes very occasionally, and usually as a result of manual
intervention to update it or make a fix, so not getting a live
copy every time shouldn't cause problems.)

This also enables us to simplify the code quite significantly.