Use the full MapIt data in tests rather than the reduced fixtures
Since we're keeping the full MapIt data for area types FED, SEN and STA
in the repository anyway now, we can remove the old reduced fixture
data and use the real data instead. This will slow down the tests
slightly, since more MapIt data is being loaded and longer pages are
rendered in the tests, but has the advantage that we're dealing with
the real data so may catch additional problems caused by data in the
full dataset. Also some of the magic numbers in the tests are clearer:
e.g. the 37 in state tests is for the 36 states + the Federal Capital
Territory.