Remove the people url since people will be loaded in the place page
Before we were implementing the page place as in the Pombola
version of the site, i.e., with dedicated pages for people and places,
hence this method was created.

When the place page was simplified, we decided to show all the
people directly in the place page, so this method is not needed
anymore.