Extract the tests for the person interface into a shared example
Since now we have two different classes to represent a person, and
both have to conform to the API that the view needs, the person
interface tests were extracted into a shared example.

The shared example extends its module with Minitest's Domain
Specific Language module, so that we have the 'it' method
available.

    http://docs.seattlerb.org/minitest/Minitest/Spec/DSL.html