Use https with a GitHub personal access token if possible
Previously this script only supported pushing to the
shineyoureye-sinatra repository over SSH. This commit changes that so
that if GITHUB_ACCESS_TOKEN is set to an appropriate personal access
token in the environment that will be used to authenticate and push over
HTTPS instead. This is important so that we can use a token to enable
pushing to the repository in the context of automated deployments.