Add methods in MembershipCSV::People that are part of the People API
The views show the current term's start date and the legislature
name if the place is a child area only.

Since MembershipCSV::People represents people that are
associated with parent areas, I am adding these methods to the
API and making them return nil, following the Null Pattern to keep
LSP working.