Fix links and images for the manually chosen featured people
Shortly we will switch to the featured people being selected by a
'featured' checkbox in prose.io, but for the moment let's just make sure
that all the links work and the images are the right ones from our image
cache. Chris Anyanwu is no longer a senator, so I replaced her by
another female senator from the APC party.