Add a method in the people finders to get a featured person
This method will return the first person it finds for a featured
summary for that legislature. So if there are no featured people,
it will return nil, and if there are more than one, it will return
one (the first).

This makes sense because there can only be one featured person
of each type at the same time in the homepage. So in order to
make one person featured, the previous will have to have their
'featured' flag set to 'false' manually in the prose.io metadata.

Another alternative would be to use the featured person whose summary
has been most recently changed according to git.