Add a link to edit person data, pointing to Nigeria data spreadsheet
At the moment, only summaries can be edited from the site. It would
be nice to edit both in the same place, but summaries and person
data are edited differently because they come from different sources:

* Summaries come from prose, so "Edit summary" points to prose's
editor-page for that person, according to the person's id.

* Person data comes from EveryPolitician, so "Edit person data" points
to a Google Spreadsheet where users can edit person data and then this
spreadsheet can be added to the sources that the scrapers consume.

The spreadsheet linked for the last item is:

    https://docs.google.com/spreadsheets/d/1YLQUsiXDcFZ4eT1PhpEQyO31-2A3lam6JTsXPACUinU/edit#gid=0