Add a class to request and parse a place's geometry data
The library used to display a map is Leaflet. Leaflet can
parse geoJSON features, so the id.geojson response from
the Mapit API will be used to get the coordinates of the
polygon that represents an area, and then that will be
passed to Leaflet so that it can be displayed in a layer
over the map.

Leaflet also needs to center the map. To get the coordinate
representing the center of the map, the geometry.json
response from the Mapit API will be used as well. This
response includes the properties "center_lat" and
"center_lon" that will be passed to Leaflet so that it
can use them to center the map in the HTML element where
it will be displayed.