Add map to place pages using Leaflet and Open Street Map
The library used to load the maps is Leaflet. At the
moment it is using Open Street Map for the tiles, since
it is open and doesn't have the restrictions that Google
Maps has.

Leaflet can parse geoJSON features, so the id.geojson
response from the Mapit API was used to get the coordinates
of the polygon that represents the area.

To calculate the center of the map, the geometry.json
response from the Mapit API was used.

Finally, since the areas vary in size, I used fitBounds()
so that all areas fit in the map.