Bugfix: place pages with no people showing empty legislature name
At the moment the place routes were doing two types of checks:

* `pass unless AREA` this prevents the code of the route from running
if the area doesn't exist and jumps to the 404 page if that is the
case.

* `pass if PERSON_TYPE.none_by_mapit_area?(AREA.id)` this was naïvely
used so that we could automatically do some sort of indirect place-type
check, so that we could jump to the right route and display the people
of the right type.

The problem with the second check is that if the place is a child area
but it happens to have no people associated with it, the app will jump
all the routes until it arrives at the state route. And the state
route uses a different class for people, since its data comes from a
CSV (because there is no data for governors in EP).

In this line of the place view:

<%= " (#{@page.legislature_name})" if @page.place.child_area? %>

the check for child area is true while there is nothing to show as
`legislature_name` because CSV people don't know their legislature name.
So it shows an empty value in the legislature name.

In this commit the app uses a single route for places, and the three
types of person in it.

Also, another problem is that the logic of the previous line in the
place view is misplaced. We should avoid logic in the views, which is
part of the reason why this bug appeared. This logic will be moved to
the place page, where it belongs, in the next commit.