Check that both a start date and a year exist in the places page
Use safe navigation operator (&.) for checks (available from Ruby 2.3 on)
instead of the longer:

    people_by_legislature.current_term_start_date.year
      if people_by_legislature.current_term_start_date

The parent place pages don't show a current term start date. The
information related with the legislature is taken from the people.
Since parent places receive an instance of people that returns nil
for the legislature related methods (legislature name and current
term start date) we have to do this check before we ask it for the
year.