Remove compass dependency since we are now using sass instead
Some time ago we moved to using sass to compile the site styles.

The compass dependency was there because we first ported the code from
the Jekyll version of the site which at the same time was ported from
the Pombola version of the site which used compass.

This dependency is now useless since we are not using compass anymore.