Add the site identifier to EP people and person classes
The person class needs to expose a slug so that we can find
a person by slug in the person page routes.

Since this site is a migration of a Pombola site, we are using
the old Pombola slugs to build the url of a person.

In person data coming from EP, these slugs are saved as
identifiers. Hence we need to pass the identifier scheme of the
site in order to find the right identifier to read the slug from.