Rewrite the event routes so that they work as the blog ones
In order to serve the site, we run the Sinatra app and then wget it
and push it to a gh-pages branch, so that it is served as a static site.

The event list page at `/info/events` works when running the app
locally, but it throws a 404 once we wget it and serve it as static.

This is because the route /info/events (the event list page) is scraped
first by wget, creating an HTML file info/events.html. Later the route
/info/events/EVENT_SLUG is scraped, and wget replaces info/events.html
with an info/events directory containing a EVENT_SLUG.html file.

In other words, the index page is saved and then deleted by wget.

To solve this, we are making the URL path for an individual event
look like /events/EVENT_SLUG instead of /info/events/EVENT_SLUG.